allow keeping database containers

+++ cloak/ci/compliance.sh
@@ -23,11 +23,13 @@ function cleanup {
 
   clean_dangling
 
-  # also stop named db containers so they don't waste dev resources
-  for container in postgres9.4 mongo3.0 mongo3.2 mongo3.4 mysql5.7 sqlserver2017; do
-    docker kill $container > /dev/null || true
-    docker rm $container > /dev/null || true
-  done
+  if [ "$KEEP_DB_CONTAINERS" != "true" ]; then
+    # also stop named db containers so they don't waste dev resources
+    for container in postgres9.4 mongo3.0 mongo3.2 mongo3.4 mysql5.7 sqlserver2017; do
+      docker kill $container > /dev/null || true
+      docker rm $container > /dev/null || true
+    done
+  fi
 
   exit $exit_status
 }
